fix some config and working in mac os

+++ src/data/simulation/init_embeddings.py
@@ -10,7 +10,7 @@ import copy
 def calc_embedding_mean(embeddings_data: dict, embedding_dimension: int=512) -> dict:
     means = {}
     for key, value in embeddings_data.items():
-        sum = np.zeros(embedding_dimension)
+        sum = np.zeros(embedding_dimension, dtype=np.float32)
         n_emb = 0
         for _, vector in value.items():
             sum += vector.numpy()
@@ -22,7 +22,7 @@ def calc_embedding_mean(embeddings_data: dict, embedding_dimension: int=512) -> 
 def calc_embedding_std(embeddings_data: dict, means: dict, embedding_dimension: int=512) -> dict:
     stds = {}
     for key, value in embeddings_data.items():
-        sum_squared_error = np.zeros(embedding_dimension)
+        sum_squared_error = np.zeros(embedding_dimension, dtype=np.float32)
         n_emb = 0
         for _, vector in value.items():
             sum_squared_error += (vector.numpy() - means[key]) ** 2
